requirements.txt
mido>=1.3
python-rtmidi>=1.5
sounddevice>=0.5
numpy>=2.0
scipy>=1.13
pyserial>=3.5
